+++ app/src/main/java/ru/gb/course1/l6recycler_hw/App.java
@@ -8,13 +8,19 @@ import ru.gb.course1.l6recycler_hw.data.CacheArticleRepositoryImpl;
 import ru.gb.course1.l6recycler_hw.domain.ArticleRepository;
 
 public class App extends Application {
-    private ArticleRepository articleRepository = new CacheArticleRepositoryImpl();
+
+    private static App sInstance = null;
+
+    public final ArticleRepository articleRepository = new CacheArticleRepositoryImpl();
 
     @Override
     public void onCreate() {
+
         super.onCreate();
+        sInstance = this;
     }
     public static App get(Context context){ return (App) context.getApplicationContext(); }
+    public static App get() {return sInstance;}
     public ArticleRepository getArticleRepo() {
         return articleRepository;
     }
