+++ app/src/main/java/ru/gb/course1/l6recycler_hw/data/CacheArticleRepositoryImpl.java
@@ -1,13 +1,10 @@
 package ru.gb.course1.l6recycler_hw.data;
 
-import android.widget.Toast;
-
 import java.util.ArrayList;
 import java.util.List;
 
 import ru.gb.course1.l6recycler_hw.domain.ArticleRepository;
 import ru.gb.course1.l6recycler_hw.domain.TimeLineEntity;
-import ru.gb.course1.l6recycler_hw.ui.ArticleEditActivity;
 
 public class CacheArticleRepositoryImpl implements ArticleRepository {
     private final List<TimeLineEntity> cache = new ArrayList<>();
