+++ app/src/main/java/ru/gb/course1/l6recycler_hw/ui/details/ArticleActivity.java
@@ -1,4 +1,4 @@
-package ru.gb.course1.l6recycler_hw.ui;
+package ru.gb.course1.l6recycler_hw.ui.details;
 
 import android.os.Bundle;
 import android.widget.Button;
