+++ app/src/main/java/ru/gb/course1/l6recycler_hw/ui/main/TimeLineViewHolder.java
@@ -1,4 +1,4 @@
-package ru.gb.course1.l6recycler_hw.ui;
+package ru.gb.course1.l6recycler_hw.ui.main;
 
 import android.view.LayoutInflater;
 import android.view.ViewGroup;
@@ -16,23 +16,27 @@ public class TimeLineViewHolder extends RecyclerView.ViewHolder {
 //    private final TextView articleTextView = itemView.findViewById(R.id.article_text_view);
     private final TextView buttonDelete = itemView.findViewById(R.id.delete_button);
     private final TextView buttonEdite = itemView.findViewById(R.id.edit_button);
-
     private OnArticleListener onArticleListener;
+    @NonNull
+    private TimeLineEntity timeLineEntity;
 
     public TimeLineViewHolder(
             @NonNull LayoutInflater inflater,
             @NonNull ViewGroup parent,
             OnArticleListener onArticleListener
     ) {
-        super(inflater.inflate(R.layout.item_article, parent, false));
+//        super(inflater.inflate(R.layout.item_article, parent, false));
+        super(LayoutInflater.from(parent.getContext()).inflate(R.layout.item_article, parent, false));
         this.onArticleListener = onArticleListener;
+        buttonDelete.setOnClickListener( v -> onArticleListener.onDeleteArticle(timeLineEntity));
+        itemView.setOnClickListener( v -> onArticleListener.onClickArticle(timeLineEntity));
+        buttonEdite.setOnClickListener( v -> onArticleListener.onEditArticle(timeLineEntity));
     }
 
     public void bind(TimeLineEntity article) {
-        buttonDelete.setOnClickListener( v -> onArticleListener.onDeleteArticle(article));
-        itemView.setOnClickListener( v -> onArticleListener.onClickArticle(article));
-        buttonEdite.setOnClickListener( v -> onArticleListener.onEditArticle(article));
+
         dateTextView.setText(article.getArticleDateTime());
         nameTextView.setText(article.getArticleTitle());
+        timeLineEntity = article;
     }
 }
