+++ app/src/main/res/layout/activity_edit_article.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".ui.MainActivity">
+    tools:context=".ui.main.MainActivity">
 
     <EditText
         android:id="@+id/activity_edit_article_title"
