+++ app/src/main/res/layout/activity_main.xml
@@ -1,22 +1,11 @@
 <?xml version="1.0" encoding="utf-8"?>
-
-<LinearLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
+<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:orientation="vertical"
-    tools:context=".ui.MainActivity">
-    <Button
-        android:id="@+id/ins_new_article_button"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_margin="8dp"
-        android:padding="8dp"
-        android:text="Добавить новую статью"
-        android:textSize="12sp" />
-    <androidx.recyclerview.widget.RecyclerView
-        android:id="@+id/recycler_view"
+    android:layout_height="match_parent">
+
+    <FrameLayout
+        android:id="@+id/activity_main__fragment_container"
         android:layout_width="match_parent"
         android:layout_height="match_parent"/>
 
-</LinearLayout>
+</FrameLayout>
